requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
python-dotenv>=1.0
openai>=1.30
pydantic>=2.6
supabase>=2.0.0
pyjwt[crypto]>=2.8.0
certifi>=2024.2.2
email-validator>=2.0.0
